+++ src/components/Schedule/Schedule.tsx
@@ -8,6 +8,8 @@ import Tooltip from '@material-ui/core/Tooltip'
 import Box from '@material-ui/core/Box/Box'
 import { ShiftOccupation } from '../../domain/Shift'
 import i18next from 'i18next'
+import { isOkWithWhite } from '../../utils/colors'
+import hexRgb from 'hex-rgb'
 // import './_materialFullCalendar.scss'
 
 class Schedule extends React.PureComponent <{
@@ -16,28 +18,34 @@ class Schedule extends React.PureComponent <{
 	lang: string
 	darkMode: boolean
 }, unknown>{
-	renderEventContent(eventInfo: {event: EventApi}): JSX.Element {
+	renderEventContent(eventInfo: {event: EventApi, backgroundColor: string}): JSX.Element {
 		const occupation = Object.values(eventInfo.event.extendedProps.occupation as ShiftOccupation)
 		const percentage = occupation[0]/occupation[1] * 100 || 0
+		const textColor = isOkWithWhite(hexRgb(eventInfo.backgroundColor)) ? 'white' : 'black'
+
 		return (
 			<Tooltip arrow title={
 				<React.Fragment>{i18next.t('schedule.shift.occupation')}: {occupation.join('/')} ({percentage.toFixed(0)}%)</React.Fragment>
 			} placement={'top'}>
-				<Box style={{height: '100%'}}>{eventInfo.event.title}
+				<Box style={{maxWidth: '100%', height: '100%', color: textColor, wordBreak: 'break-word', overflowY: 'hidden'}}>
+					<p style={{margin: '0'}}>{eventInfo.event.title}</p>
+					<p style={{margin: '0'}}>{eventInfo.event.extendedProps.campus}</p>
 				</Box>
 			</Tooltip>
 		)
 	}
 
 	onEventClick(info: EventClickArg): void {
-		const withCtrl = info.jsEvent.ctrlKey
-		if (withCtrl) {
-			const url = info.event.extendedProps.courseUrl
-			const win = window.open(url)
-			win?.focus()
-		} else {
-			this.props.onSelectedEvent(info.event.id)	
-		}
+		// TODO: The courses don't have url for now
+		// const withCtrl = info.jsEvent.ctrlKey
+		// if (withCtrl) {
+		// 	const url = info.event.extendedProps.courseUrl
+		// 	const win = window.open(url)
+		// 	win?.focus()
+		// } else {
+		// 	this.props.onSelectedEvent(info.event.id)	
+		// }
+		this.props.onSelectedEvent(info.event.id)	
 	}
 
 	render(): React.ReactNode {